add pmg_verify_tls_policy_strict and use it in API
This patch splits the parsing of tls_policies in 2 parts:
While reading we just require a line to start with one of the valid tls_policies,
while writing we only accept one of the policies w/o any attributes.
This should help users, who already have a manually crafted file in place, to
use API-calls for adding/modifying entries.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>