add rule's score to pmg-smtp-filter logline
We already log which Spamassassin rules apply to a mail. Given that the scores
depend on configuration and setup (e.g. AWL) writing them in the log provides
a quick overview of Spamassassin performance, and spares admins and support from
having to gather the complete mail just for assessing Spamassessin.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>