fix #2172: sort mynetworks template var to make postfix config rewrite stable
The mynetworks template_var is written to postfix/main.cf, causing a
postfix restart on every change. Since mynetworks is a hash the order
of the networks potentially changes with every invocation. This shows
quite readily in clustered setups where pmgmirror writes the configs
and checks for changes once every 2 minutes.

fixes #2172

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>